+++ demo/united.R
@@ -4,7 +4,7 @@ library(rvest)
 united <- html_session("http://www.united.com/")
 
 login <- united %>%
-  html_node("form[name=LoginForm]") %>%
+  html_element("form[name=LoginForm]") %>%
   html_form() %>%
   html_form_set(
     MpNumber = "GY797363",
@@ -15,6 +15,6 @@ logged_in <- united %>% session_submit(login)
 
 logged_in %>%
   follow_link("View account") %>%
-  html_node("#ctl00_ContentInfo_AccountSummary_spanEliteMilesNew") %>%
+  html_element("#ctl00_ContentInfo_AccountSummary_spanEliteMilesNew") %>%
   html_text() %>%
   readr::parse_number()