+++ tests/testthat/test-rename.R
@@ -6,6 +6,13 @@ test_that("xml functions are deprecated", {
   expect_snapshot(. <- xml_nodes(x, "p"))
 })
 
+test_that("html_node(s) is superseded (no warnings)", {
+  x <- minimal_html("<p>Hello</p>")
+
+  expect_equal(html_node(x, "p"), html_element(x, "p"))
+  expect_equal(html_nodes(x, "p"), html_elements(x, "p"))
+})
+
 test_that("set_values() is deprecated", {
   html <- minimal_html('
     <form><input type="text" name="text" /></form>